MeshSinusoid
This effect causes the object to resemble a sinusoid in its (x,y) plane.
The intermediate values were nondimensionalised with respect to both the
object's height and width. Nondimensionalisation allows: 1) the object
to resemble one full sinusoidal cycle, and 2) the function to take as
its argument the percentage of change with respect to the object's width
that one would like the object to incur.